refactor: common 폴더 내 컴포넌트 경로 수정 [TRL-199]
