+++ src/mocks/handlers.ts
@@ -192,7 +192,7 @@ const getUserInfo = rest.get('/api/user-info', async (req, res, ctx) => {
   return res(
     ctx.json({
       totalDistanceOfPastTrip: 410,
-      totalNumOfTrip: 10,
+      totalNumOfTripPlan: 10,
       badgeImgUrl:
         'https://user-images.githubusercontent.com/84956036/227441024-9853dda6-2100-466a-af20-b13d2e720f5f.png',
     })
