+++ src/pages/Login/index.tsx
@@ -4,9 +4,10 @@ import styled from 'styled-components';
 
 import Button from '@/components/common/Button';
 import Description from '@/components/common/Description';
+import Flex from '@/components/common/Flex';
 import Line from '@/components/common/Line';
 import Logo from '@/components/common/Logo';
-import Spacer from '@/components/common/Spacer';
+import Spacing from '@/components/common/Spacing';
 import OAuthBtnContainer from '@/components/OAuthBtnContainer';
 import color from '@/constants/color';
 import UserStatus, { UserStatusTypes } from '@/states/userStatus';
@@ -24,19 +25,21 @@ const Login = () => {
   return (
     <Layout>
       <Logo width={76} height={50} />
-      <Spacer top={85} bottom={35}>
-        <Description color={color.gray2} fontSize={1.4}>
-          SNS 간편 로그인
-        </Description>
-      </Spacer>
+      <Spacing height={85} />
+      <Description color={color.gray2} fontSize={1.4}>
+        SNS 간편 로그인
+      </Description>
+      <Spacing height={35} />
       <OAuthBtnContainer />
-      <Spacer top={35} bottom={35}>
+      <Spacing height={35} />
+      <Flex alignCenter>
         <Line width={193} right={35} />
         <Description color={color.gray2} fontSize={1.4}>
           또는
         </Description>
         <Line width={193} left={35} />
-      </Spacer>
+      </Flex>
+      <Spacing height={35} />
       <Button type="button" onClick={handleVisitorBtnClick} btnSize="large">
         로그인 없이 둘러보기
       </Button>
